Add the first non-universal logic:  A preprocessor that converts incoming updates into a simplified Context object.
A followup will introduce a dispatcher that uses this preprocessor and then dispatches Context objects (rather than raw Telegram Update objects) to its registered handlers.

This is not 100%:  Only three types of updates are supported, and for example InlineQuery updates include an offset parameter that should be preserved but isn't. But it's a good base to begin work on the dispatcher.