Identify forwarded messages based on the presence of the `forward_date` field, which catches forwarded channel posts, forwarded private messages, and forwarded anonymous messages.
The `Context.forward_from` remains unchanged (except in the unlikely event of a message that contains `forward_from` but not `forward_date`). A new `Context.forwarded` marks whether the message is forwarded or not.

For now, `Context.text` remains completely unset for all forwarded messages, but this may change in a followup (see nmlorg/metabot#77).