Support https://core.telegram.org/bots/api#sendmessage.
This is slightly hairy—I don't particularly like the style of raising one type of exception and then converting it to a different kind of exception (Telegram.ChatNotFound() -> BotAPI.Error(400, ...)). At minimum, this logic is going to need to be moved to a separate BotAPI.get_chat_view (so it can be reused for BotAPI.api_sendphoto, etc.).

The mess with _Chat and _ChatView is necessary because messages in private chats include the current `about` info of the other participant (so when bot1234 sends a message to user1000, it gets user1000's current first_name, etc. returned in the message's `chat` field).

The `if chatid < 0:` section of Telegram.get_chat_view is currently unused, but shows what I'm thinking for supporting group chats.